+++ STM32CrapSynthFirmware/src/spi.h
@@ -12,7 +12,7 @@
 
 void spi_init();
 void spi1_send_via_dma(uint8_t* sendbuf, uint16_t size);
-void spi2_send_via_dma(uint8_t* sendbuf, uint16_t size);
+void spi2_send_via_dma(uint16_t* sendbuf, uint16_t size);
 void spi1_receive_via_dma(uint8_t* sendbuf, uint8_t* receivebuf, uint16_t size);
 
 #endif /* SPI_H_ */
